@@ -31,7 +31,7 @@ setup(
     author='JAX team',
     author_email='jax-dev@google.com',
     packages=['jaxlib'],
-    python_requires='>=3.5',
+    python_requires='>=3.6',
     install_requires=['scipy', 'numpy>=1.12', 'absl-py'],
     url='https://github.com/google/jax',
     license='Apache-2.0',
