@@ -13,11 +13,11 @@
 # limitations under the License.
 
 
-from collections import namedtuple, defaultdict
+from collections import defaultdict
 from contextlib import contextmanager
-import itertools as it
 import operator as op
 import threading
+from typing import Any, Callable, Dict, List, Sequence, Set, Type
 
 from absl import logging
 import numpy as onp
@@ -28,10 +28,9 @@ from .. import linear_util as lu
 from .. import lazy
 from ..abstract_arrays import (ConcreteArray, ShapedArray, array_types,
                                raise_to_shaped)
-from ..util import (partial, unzip2, concatenate, prod, safe_map,
+from ..util import (partial, unzip2, prod, safe_map,
                     extend_name_stack, wrap_name)
 from ..lib import xla_bridge as xb
-from .xla import aval_to_xla_shape, xla_destructure
 from .batching import broadcast, not_mapped
 from . import batching
 from . import partial_eval as pe
@@ -47,11 +46,10 @@ _map = safe_map
 
 def identity(x): return x
 
-def shard_args(backend, devices, assignments, axis_size, tuple_args, args):
+def shard_args(devices, assignments, axis_size, args):
   """Shard each argument data array along its leading axis.
 
   Args:
-    backend: the platform to be used
     devices: list of Devices mapping replica index to a physical device.
     assignments: list of integers with the same length as `devices` mapping
       replica index to an index along the leading axis (i.e. a shard).
@@ -98,14 +96,10 @@ def shard_args(backend, devices, assignments, axis_size, tuple_args, args):
       for r, buf in enumerate(bufs):
         buffers[r][a] = buf
 
-  if tuple_args:
-    buffers = [[xla.make_tuple(bufs, devices[r], backend)]
-               for r, bufs in enumerate(buffers)]
-
   return buffers
 
 
-shard_arg_handlers = {}
+shard_arg_handlers: Dict[Any, Callable[[Any, Any, Any], Sequence[Any]]] = {}
 shard_arg_handlers[core.Unit] = \
     lambda x, devices, _: [xla.device_put(core.unit, d) for d in devices]
 def _shard_array(x, devices, assignments):
@@ -124,11 +118,14 @@ shard_arg_handlers[xla.DeviceArray] = _shard_device_array
 def shard_aval(size, aval):
   try:
     return shard_aval_handlers[type(aval)](size, aval)
-  except KeyError:
-    raise TypeError("No shard_aval handler for type: {}".format(type(aval)))
-shard_aval_handlers = {}
+  except KeyError as err:
+    raise TypeError("No shard_aval handler for type: {}".format(type(aval))
+                    ) from err
+shard_aval_handlers: Dict[Type[core.AbstractValue], Callable[[int, Any], Any]] = {}
 shard_aval_handlers[core.AbstractUnit] = lambda size, x: x
 def _shard_abstract_array(size, x):
+  if not x.shape:
+    raise ValueError("Scalar cannot be split across {} shards.".format(size))
   if x.shape[0] != size:
     raise ValueError("Axis size {} does not match leading dimension of "
                      "shape {}".format(size, x.shape))
@@ -138,9 +135,11 @@ shard_aval_handlers[ShapedArray] = _shard_abstract_array
 def aval_to_result_handler(size, nrep, aval):
   try:
     return pxla_result_handlers[type(aval)](size, nrep, aval)
-  except KeyError:
-    raise TypeError("No pxla_result_handler for type: {}".format(type(aval)))
-pxla_result_handlers = {}
+  except KeyError as err:
+    raise TypeError("No pxla_result_handler for type: {}".format(type(aval))
+                    ) from err
+PxlaResultHandler = Callable[..., Callable[[Any], Any]]
+pxla_result_handlers: Dict[Type[core.AbstractValue], PxlaResultHandler] = {}
 pxla_result_handlers[core.AbstractUnit] = lambda *_: lambda _: core.unit
 def array_result_handler(size, nrep, aval):
   full_aval = ShapedArray((size,) + aval.shape, aval.dtype)
@@ -246,10 +245,47 @@ def apply_parallel_primitive(prim, *args, **params):
     shape = (logical_size(dynamic_axis_env[axis_name]),)
   return parallel_pure_rules[prim](*args, shape=shape, **params)
 
-parallel_pure_rules = {}
+parallel_pure_rules: Dict[core.Primitive, Callable] = {}
 
 
 def axis_index(axis_name):
+  """Return the index along the pmapped axis ``axis_name``.
+
+  Args:
+    axis_name: hashable Python object used to name the pmapped axis (see the
+      ``pmap`` docstring for more details).
+
+  Returns:
+    An integer representing the index.
+
+  For example, with 8 XLA devices available:
+
+  >>> from functools import partial
+  >>> @partial(pmap, axis_name='i')
+  ... def f(_):
+  ...   return lax.axis_index('i')
+  ...
+  >>> f(np.zeros(4))
+  ShardedDeviceArray([0, 1, 2, 3], dtype=int32)
+  >>> f(np.zeros(8))
+  ShardedDeviceArray([0, 1, 2, 3, 4, 5, 6, 7], dtype=int32)
+  >>> @partial(pmap, axis_name='i')
+  ... @partial(pmap, axis_name='j')
+  ... def f(_):
+  ...   return lax.axis_index('i'), lax.axis_index('j')
+  ...
+  >>> x, y = f(np.zeros((4, 2)))
+  >>> print(x)
+  [[0 0]
+   [1 1]
+   [2 2]
+   [3 3]]
+  >>> print(y)
+  [[0 1]
+   [0 1]
+   [0 1]
+   [0 1]]
+  """
   dynamic_axis_env = _thread_local_state.dynamic_axis_env
   frame = dynamic_axis_env[axis_name]
   sizes = dynamic_axis_env.sizes[:dynamic_axis_env.index(frame)+1]
@@ -266,7 +302,7 @@ def _axis_index_partial_eval(trace, _, **params):
   # during pmap lowering. It is like the standard JaxprTrace.process_primitive
   # rule except that we don't attempt to lower out of the trace.
   out_aval = ShapedArray((), onp.int32)
-  out_tracer = pe.JaxprTracer(trace, pe.PartialVal((out_aval, core.unit)), None)
+  out_tracer = pe.JaxprTracer(trace, pe.PartialVal.unknown(out_aval), None)
   eqn = pe.new_eqn_recipe([], [out_tracer], axis_index_p, params)
   out_tracer.recipe = eqn
   return out_tracer
@@ -381,7 +417,7 @@ xla.canonicalize_dtype_handlers[ShardedDeviceArray] = identity
 
 
 class ChunkedDeviceArray(ShardedDeviceArray):
-  __slots__ = []
+  __slots__: List[str] = []
   _collect = staticmethod(onp.concatenate)
 
   def __init__(self, axis_size, aval, device_buffers):
@@ -401,7 +437,7 @@ xla.canonicalize_dtype_handlers[ChunkedDeviceArray] = identity
 
 ### the xla_pmap primitive and its rules are comparable to xla_call in xla.py
 
-def xla_pmap_impl(fun, *args, backend, axis_name, axis_size, global_axis_size,
+def xla_pmap_impl(fun: lu.WrappedFun, *args, backend, axis_name, axis_size, global_axis_size,
                   devices, name, mapped_invars=None):
   abstract_args = map(xla.abstractify, args)
   compiled_fun = parallel_callable(fun, backend, axis_name, axis_size,
@@ -457,15 +493,13 @@ def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
       return fun.call_wrapped(*args)
 
   avals = tuple(map(partial(shard_aval, axis_size), avals))
-  pvals = [pe.PartialVal((aval, core.unit)) for aval in avals]
+  pvals = [pe.PartialVal.unknown(aval) for aval in avals]
   # We add a dummy first invar, to carry the trace details to `dynamic_fun`
-  pval = pe.PartialVal([core.abstract_unit, core.unit])  # dummy value for axis env
-  with core.new_master(pe.StagingJaxprTrace, True) as master:
-    jaxpr, (out_pvals, consts, env) = pe.trace_to_subjaxpr(
-        dynamic_fun, master, False).call_wrapped([pval] + pvals)
-    jaxpr.invars = jaxpr.invars[1:]  # ignore dummy
-    assert not env
-    del master
+  pval = pe.PartialVal.unknown(core.abstract_unit)  # dummy value for axis env
+  jaxpr, out_pvals, consts = pe.trace_to_jaxpr(
+      dynamic_fun, [pval] + pvals, instantiate=False, stage_out=True, bottom=True)
+  jaxpr.invars = jaxpr.invars[1:]  # ignore dummy
+
   out_pvs, out_consts = unzip2(out_pvals)
 
   # TODO(skye,mattjj): allow more collectives on multi-host as we test them, but
@@ -539,22 +573,24 @@ def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
                        % (num_global_replicas, len(devices)))
 
   device_assignment = tuple(d.id for d in devices)
-  compiled = built.Compile(
-      compile_options=xb.get_compile_options(
+  compile_options = xb.get_compile_options(
           num_replicas=num_global_replicas,
           num_partitions=1,
-          device_assignment=device_assignment),
-      backend=xb.get_backend(backend))
+          device_assignment=device_assignment)
+  compile_options.tuple_arguments = tuple_args
+  compiled = built.Compile(compile_options=compile_options,
+                           backend=xb.get_backend(backend))
 
-  handle_args = partial(shard_args, backend, compiled.local_devices(),
+  handle_args = partial(shard_args, compiled.local_devices(),
                         assign_shards_to_replicas(num_local_replicas, axis_size),
-                        axis_size, tuple_args)
+                        axis_size)
   handle_outs = _pvals_to_results_handler(axis_size, num_local_replicas,
                                           out_pvals, compiled.local_devices(),
                                           backend)
-  return partial(execute_replicated, compiled, backend, handle_args, handle_outs)
+  return partial(execute_replicated, compiled, backend, handle_args,
+                 handle_outs)
 
-multi_host_supported_collectives = set()
+multi_host_supported_collectives: Set[core.Primitive] = set()
 
 class ResultToPopulate(object): pass
 result_to_populate = ResultToPopulate()
@@ -566,7 +602,7 @@ def _pvals_to_results_handler(size, nrep, out_pvals, devices, backend):
   def handler(out_bufs):
     buffers = [[result_to_populate] * nrep for _ in range(nouts)]
     for r, tuple_buf in enumerate(out_bufs):
-      for i, buf in enumerate(tuple_buf.destructure()):
+      for i, buf in enumerate(tuple_buf):
         buffers[i][r] = buf
     assert not any(buf is result_to_populate for bufs in buffers
                    for buf in bufs)
@@ -605,7 +641,7 @@ def replicate(val, axis_size, nrep, devices=None, backend=None):
     devices = xb.get_backend(backend).get_default_device_assignment(nrep)
   assert nrep == len(devices)
 
-  aval = xla.abstractify(val)
+  aval = xla.abstractify(val)  # type: ShapedArray
   aval = ShapedArray((axis_size,) + aval.shape, aval.dtype)
   device_buffers = [xla.device_put(val, d) for d in devices]
   return ShardedDeviceArray(aval, device_buffers)
@@ -635,7 +671,7 @@ def _pval_to_result_handler(axis_size, nrep, pval, devices, backend):
 
 def execute_replicated(compiled, backend, in_handler, out_handler, *args):
   input_bufs = in_handler(args)
-  out_bufs = compiled.ExecutePerReplica(list(input_bufs))
+  out_bufs = compiled.ExecuteOnLocalDevices(list(input_bufs))
   return out_handler(out_bufs)
 
 
@@ -649,8 +685,7 @@ xla_pmap_p.def_impl(xla_pmap_impl)
 def _pmap_translation_rule(c, axis_env,
                            in_nodes, name_stack, axis_name, axis_size,
                            global_axis_size, devices, name,
-                           call_jaxpr, backend=None,
-                           mapped_invars=None):
+                           call_jaxpr, *, backend=None, mapped_invars):
   # We in-line here rather than generating a Call HLO as in the xla_call
   # translation rule just because the extra tuple stuff is a pain.
   if axis_env.devices is not None or (axis_env.names and devices is not None):
@@ -659,45 +694,61 @@ def _pmap_translation_rule(c, axis_env,
     global_axis_size = axis_size
   new_env = xla.extend_axis_env(axis_env, axis_name, global_axis_size)
   # Shard the in_nodes that are mapped
+  in_avals = [v.aval for v in call_jaxpr.invars]
   in_nodes_sharded = (
-    _xla_shard(c, new_env, in_node) if in_node_mapped else in_node
-    for in_node, in_node_mapped in zip(in_nodes, mapped_invars))
+    _xla_shard(c, aval, new_env, in_node) if in_node_mapped else in_node
+    for aval, in_node, in_node_mapped in zip(in_avals, in_nodes, mapped_invars))
 
   sharded_outs = xla.jaxpr_subcomp(
       c, call_jaxpr, backend, new_env, (),
       extend_name_stack(name_stack, wrap_name(name, 'pmap')), *in_nodes_sharded)
-  outs = [_xla_unshard(c, new_env, shard) for shard in sharded_outs]
+  out_avals = [v.aval for v in call_jaxpr.outvars]
+  outs = [_xla_unshard(c, aval, new_env, shard, backend=backend)
+          for aval, shard in zip(out_avals, sharded_outs)]
   return c.Tuple(*outs)
 
 xla.call_translations[xla_pmap_p] = _pmap_translation_rule
 ad.primitive_transposes[xla_pmap_p] = partial(ad.map_transpose, xla_pmap_p)
 pe.map_primitives.add(xla_pmap_p)
 
-def _xla_shard(c, axis_env, x):
-  xla_shape = c.GetShape(x)
-  if xla_shape.is_tuple():
-    assert not xla_shape.tuple_shapes()
+def _xla_shard(c, aval, axis_env, x):
+  if aval is core.abstract_unit:
     return x
-  else:
-    dims = list(xla_shape.dimensions())
+  elif isinstance(aval, ShapedArray):
+    dims = list(c.GetShape(x).dimensions())
     zero = c.Constant(onp.zeros((), dtype=onp.uint32))
     idxs = [_unravel_index(c, axis_env)] + [zero] * (len(dims) - 1)
     return c.Reshape(c.DynamicSlice(x, idxs, [1] + dims[1:]), None, dims[1:])
+  else:
+    raise TypeError((aval, c.GetShape(x)))
 
 # TODO(b/110096942): more efficient gather
-def _xla_unshard(c, axis_env, x):
-  xla_shape = c.GetShape(x)
-  if xla_shape.is_tuple():
-    assert not xla_shape.tuple_shapes()
+def _xla_unshard(c, aval, axis_env, x, backend):
+  if aval is core.abstract_unit:
     return x
-  else:
+  elif isinstance(aval, ShapedArray):
+    # TODO(mattjj): remove this logic when AllReduce PRED supported on CPU / GPU
+    convert_bool = (onp.issubdtype(aval.dtype, onp.bool_)
+                    and xb.get_backend(backend).platform in ('cpu', 'gpu'))
+    if convert_bool:
+      x = c.ConvertElementType(x, xb.dtype_to_etype(onp.float32))
+
+    xla_shape = c.GetShape(x)
     dims = list(xla_shape.dimensions())
     padded = c.Broadcast(c.Constant(onp.array(0, xla_shape.numpy_dtype())),
                          [axis_env.sizes[-1]] + dims)
     zero = c.Constant(onp.zeros((), dtype=onp.uint32))
     idxs = [_unravel_index(c, axis_env)] + [zero] * len(dims)
     padded = c.DynamicUpdateSlice(padded, c.Reshape(x, None, [1] + dims), idxs)
-    return c.CrossReplicaSum(padded, xla.axis_groups(axis_env, axis_env.names[-1]))
+    out = c.CrossReplicaSum(padded, xla.axis_groups(axis_env, axis_env.names[-1]))
+
+    # TODO(mattjj): remove this logic when AllReduce PRED supported on CPU / GPU
+    if convert_bool:
+      nonzero = c.Ne(out, c.Constant(onp.array(0, dtype=onp.float32)))
+      out = c.ConvertElementType(nonzero, xb.dtype_to_etype(onp.bool_))
+    return out
+  else:
+    raise TypeError((aval, c.GetShape(x)))
 
 def _unravel_index(c, axis_env):
   div = c.Constant(onp.array(axis_env.nreps // prod(axis_env.sizes), onp.uint32))
@@ -796,17 +847,23 @@ class SplitAxisTrace(core.Trace):
           # if the name matches this tracer's name, apply the split_axis rule
           try:
             rule = split_axis_rules[primitive]
-          except KeyError:
+          except KeyError as err:
             msg = "split_axis for {} not implemented. Open a feature request!"
-            raise NotImplementedError(msg.format(primitive))
+            raise NotImplementedError(msg.format(primitive)) from err
           which_mapped = [n is not not_mapped for n in names_in]
           val_out, is_mapped = rule(vals_in, which_mapped, **params)
           name_out = name if is_mapped else not_mapped
-          return SplitAxisTracer(self, name_out, val_out)
+          if primitive.multiple_results:
+            return [SplitAxisTracer(self, name_out, v) for v in val_out]
+          else:
+            return SplitAxisTracer(self, name_out, val_out)
         else:
           # if not, bind the primitive without any processing
           val_out = primitive.bind(*vals_in, **params)
-          return SplitAxisTracer(self, name, val_out)
+          if primitive.multiple_results:
+            return [SplitAxisTracer(self, name, v) for v in val_out]
+          else:
+            return SplitAxisTracer(self, name, val_out)
       else:
         # if it's not a pmap collective primitive, act just like batching
         rule = batching.get_primitive_batcher(primitive)
@@ -822,7 +879,7 @@ class SplitAxisTrace(core.Trace):
         else:
           return new_tracer(val_out, axis_out)
 
-  def process_call(self, call_primitive, f, tracers, params):
+  def process_call(self, call_primitive, f: lu.WrappedFun, tracers, params):
     assert call_primitive.multiple_results
     if call_primitive in pe.map_primitives:
       return self.process_map(call_primitive, f, tracers, params)
@@ -835,7 +892,7 @@ class SplitAxisTrace(core.Trace):
         vals_out = call_primitive.bind(f, *vals, **params)
         return [SplitAxisTracer(self, a, x) for a, x in zip(names_out(), vals_out)]
 
-  def process_map(self, map_primitive, f, tracers, params):
+  def process_map(self, map_primitive, f: lu.WrappedFun, tracers, params):
     vals, names = unzip2((t.val, t.axis_name) for t in tracers)
     if all(name is not_mapped for name in names):
       return map_primitive.bind(f, *vals, **params)
@@ -860,4 +917,4 @@ class SplitAxisTrace(core.Trace):
     return  val, todo
 
 
-split_axis_rules = {}
+split_axis_rules: Dict[core.Primitive, Callable] = {}
