@@ -14,15 +14,12 @@
 # limitations under the License.
 
 
-from functools import partial
 import numpy as onp
 
 from jax.numpy import lax_numpy as np
 from jax.numpy.vectorize import vectorize
 from jax import ad_util
 from jax import api
-from jax import api_util
-from jax import core
 from jax import lax
 from jax import ops
 from jax import dtypes
@@ -34,7 +31,6 @@ from jax.abstract_arrays import ShapedArray
 from jax.core import Primitive
 from jax.lax import (standard_primitive, standard_unop, naryop_dtype_rule,
                      _float, _complex, _input_dtype, _broadcasting_select)
-from jax.lib import xla_client
 from jax.lib import lapack
 from jax.lib import cusolver
 
@@ -74,9 +70,15 @@ def svd(x, full_matrices=True, compute_uv=True):
 def triangular_solve(a, b, left_side=False, lower=False, transpose_a=False,
                      conjugate_a=False, unit_diagonal=False):
   conjugate_a = conjugate_a and np.issubdtype(lax.dtype(a), np.complexfloating)
-  return triangular_solve_p.bind(
+  singleton = np.ndim(b) == np.ndim(a) - 1
+  if singleton:
+    b = np.expand_dims(b, -1 if left_side else -2)
+  out = triangular_solve_p.bind(
       a, b, left_side=left_side, lower=lower, transpose_a=transpose_a,
       conjugate_a=conjugate_a, unit_diagonal=unit_diagonal)
+  if singleton:
+    out = out[..., 0] if left_side else out[..., 0, :]
+  return out
 
 
 # utilities
@@ -360,7 +362,7 @@ def triangular_solve_transpose_rule(
     unit_diagonal):
   # Triangular solve is nonlinear in its first argument and linear in its second
   # argument, analogous to `div` but swapped.
-  assert a is not ad.undefined_primal and b is ad.undefined_primal
+  assert not ad.is_undefined_primal(a) and ad.is_undefined_primal(b)
   if cotangent is ad_util.zero:
     cotangent_b = ad_util.zero
   else:
@@ -494,7 +496,7 @@ def _lu_unblocked(a):
   return lax.fori_loop(0, min(m, n), body, (pivot, perm, a))
 
 
-def _lu_blocked(a, block_size=32):
+def _lu_blocked(a, block_size=128):
   """Blocked LU decomposition, as an unrolled loop."""
   m, n = a.shape
   r = min(m, n)
@@ -502,13 +504,12 @@ def _lu_blocked(a, block_size=32):
   for k in range(0, r, block_size):
     b = min(r - k, block_size)
     block_pivot, perm, lu_block = _lu_unblocked(a[k:, k:k+b])
-    a = ops.index_update(a, ops.index[k:, k:k+b], lu_block)
 
-    a = ops.index_update(a, ops.index[k:, :k], a[perm + k, :k])
+    a = ops.index_update(a, ops.index[k:, :], a[perm + k, :])
+    a = ops.index_update(a, ops.index[k:, k:k+b], lu_block)
     pivot = ops.index_update(pivot, ops.index[k:k+b], block_pivot + k)
 
     if k + b < n:
-      a = ops.index_update(a, ops.index[k:, k+b:], a[perm + k, k+b:])
       a = ops.index_update(
         a, ops.index[k:k+b, k+b:],
         triangular_solve(a[k:k+b, k:k+b], a[k:k+b, k+b:],
@@ -619,7 +620,7 @@ lu_p = Primitive('lu')
 lu_p.multiple_results = True
 lu_p.def_impl(_lu_impl)
 lu_p.def_abstract_eval(_lu_abstract_eval)
-xla.translations[lu_p] = xla.lower_fun(_lu_python, instantiate=True)
+xla.translations[lu_p] = xla.lower_fun(_lu_python)
 ad.primitive_jvps[lu_p] = _lu_jvp_rule
 batching.primitive_batchers[lu_p] = _lu_batching_rule
 
