@@ -62,7 +62,7 @@ dynamic positional arguments for the generators, and also the auxiliary output
 data must be immutable, because it will be stored in function memoization tables.
 """
 
-
+from typing import Any, Tuple
 import weakref
 
 from .util import curry
@@ -81,7 +81,8 @@ class Store(object):
     self._val = _EMPTY_STORE_VALUE
 
   def store(self, val):
-    assert self._val is _EMPTY_STORE_VALUE, "Store occupied"
+    if self._val is not _EMPTY_STORE_VALUE:
+      raise StoreException("Store occupied")
     self._val = val
 
   @property
@@ -122,7 +123,7 @@ class WrappedFun(object):
   def __name__(self):
     return getattr(self.f, '__name__', '<unnamed wrapped function>')
 
-  def wrap(self, gen, gen_static_args, out_store):
+  def wrap(self, gen, gen_static_args, out_store) -> 'WrappedFun':
     """Add another transform and its store."""
     return WrappedFun(self.f, ((gen, gen_static_args),) + self.transforms,
                       (out_store,) + self.stores, self.params)
@@ -172,7 +173,7 @@ class WrappedFun(object):
             self.params == other.params)
 
 @curry
-def transformation(gen, fun: WrappedFun, *gen_static_args):
+def transformation(gen, fun: WrappedFun, *gen_static_args) -> WrappedFun:
   """Adds one more transformation to a WrappedFun.
   Args:
     gen: the transformation generator function
@@ -182,7 +183,7 @@ def transformation(gen, fun: WrappedFun, *gen_static_args):
   return fun.wrap(gen, gen_static_args, None)
 
 @curry
-def transformation_with_aux(gen, fun: WrappedFun, *gen_static_args):
+def transformation_with_aux(gen, fun: WrappedFun, *gen_static_args) -> Tuple[WrappedFun, Any]:
   """Adds one more transformation with auxiliary output to a WrappedFun."""
   out_store = Store()
   out_thunk = lambda: out_store.val
@@ -194,7 +195,7 @@ def fun_name(f):
   except:
     return str(f)
 
-def wrap_init(f, params={}):
+def wrap_init(f, params={}) -> WrappedFun:
   """Wraps function `f` as a `WrappedFun`, suitable for transformation."""
   return WrappedFun(f, (), (), tuple(sorted(params.items())))
 
@@ -209,7 +210,7 @@ def cache(call):
   """
   fun_caches = weakref.WeakKeyDictionary()
 
-  def memoized_fun(fun, *args):
+  def memoized_fun(fun: WrappedFun, *args):
     cache = fun_caches.setdefault(fun.f, {})
     key = (fun.transforms, fun.params, args)
     result = cache.get(key, None)
@@ -228,3 +229,24 @@ def cache(call):
 def hashable_partial(x, *args):
   ans = yield (x,) + args, {}
   yield ans
+
+
+def merge_linear_aux(aux1, aux2):
+  try:
+    out1 = aux1()
+  except StoreException:
+    # store 1 was not occupied, so store 2 better be
+    try:
+      out2 = aux2()
+    except StoreException:
+      raise StoreException("neither store occupied")
+    else:
+      return False, out2
+  else:
+    # store 1 was occupied, so let's check store 2 is not occupied
+    try:
+      out2 = aux2()
+    except StoreException:
+      return True, out1
+    else:
+      raise StoreException("both stores occupied")
