@@ -13,24 +13,23 @@
 # limitations under the License.
 
 
-import numpy as onp
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy import lax_numpy as np
+from ...numpy import lax_numpy as jnp
 from ..special import xlogy, xlog1py
 
 
-@np._wraps(osp_stats.bernoulli.logpmf, update_doc=False)
+@jnp._wraps(osp_stats.bernoulli.logpmf, update_doc=False)
 def logpmf(k, p, loc=0):
-  k, p, loc = np._promote_args_inexact("bernoulli.logpmf", k, p, loc)
-  zero = np._constant_like(k, 0)
-  one = np._constant_like(k, 1)
+  k, p, loc = jnp._promote_args_inexact("bernoulli.logpmf", k, p, loc)
+  zero = jnp._constant_like(k, 0)
+  one = jnp._constant_like(k, 1)
   x = lax.sub(k, loc)
   log_probs = xlogy(x, p) + xlog1py(lax.sub(one, x), -p)
-  return np.where(np.logical_or(lax.lt(x, zero), lax.gt(x, one)),
-                  -np.inf, log_probs)
+  return jnp.where(jnp.logical_or(lax.lt(x, zero), lax.gt(x, one)),
+                  -jnp.inf, log_probs)
 
-@np._wraps(osp_stats.bernoulli.pmf, update_doc=False)
+@jnp._wraps(osp_stats.bernoulli.pmf, update_doc=False)
 def pmf(k, p, loc=0):
-  return np.exp(logpmf(k, p, loc))
+  return jnp.exp(logpmf(k, p, loc))
